+++ telegram-bot-core/src/main/kotlin/io/github/dehuckakpyt/telegrambot/model/telegram/internal/GetUpdates.kt
@@ -2,17 +2,18 @@ package io.github.dehuckakpyt.telegrambot.model.telegram.`internal`
 
 import com.fasterxml.jackson.`annotation`.JsonProperty
 import kotlin.Int
+import kotlin.Long
 import kotlin.String
 import kotlin.collections.Iterable
 
 /**
- * Created on 02.06.2024.
+ * Created on 03.06.2024.
  *
  * @author KScript
  */
 internal data class GetUpdates(
     @get:JsonProperty("offset")
-    public val offset: Int? = null,
+    public val offset: Long? = null,
     @get:JsonProperty("limit")
     public val limit: Int? = null,
     @get:JsonProperty("timeout")
